# specs/001-operational-mvp/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CashLens Operational MVP contract delta
  version: 0.2.0
  description: >-
    Design contract for production-readiness changes. Existing documented routes remain;
    implementation must merge these additions into generated Swagger.

    Envelope rule: every 2xx response body is the existing global BaseResponseDto envelope
    produced by the APP_INTERCEPTOR-registered BaseResponseInterceptor:
    {success, data, message, timestamp} (+ additive correlationId from T064/P2.1).
    Payload schemas below describe only `data`; each 2xx response composes them with `Envelope`
    via allOf. Error responses are not enveloped (Nest exception shape + additive code/correlationId/fields).

    Compatibility rule: existing field names and enum values are retained; new fields are additive.
    Intentional changes are listed in plan.md "API Compatibility".

    Deployment rule (OPS-009): in production the web app and this API share one public origin;
    the proxy routes /api/ to the API and / to the web app, and the web app calls this API by the relative path /api.

    Authorization rule (SEC-005, US1 closure): 401 for a missing, invalid, or inactive-account session; on
    administrator-only operations authorization precedes body validation, so a non-administrator gets 403 whatever
    fields a well-formed JSON body contains; owner-scoped operations answer absent, not-owned, and malformed ids with
...
servers:
  - url: /api
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
  schemas:
    Envelope:
      description: Existing BaseResponseDto shape (apps/api/src/common/dto/base-response.dto.ts).
      type: object
      required: [success, data, message, timestamp]
      properties:
        success: { type: boolean, const: true }
        data: {}
        message: { type: string, example: Success }
        timestamp: { type: string, format: date-time }
        correlationId: { type: string, description: Additive (T064); absent before P2.1 lands }
    Error:
      description: Nest exception shape retained; code/correlationId/fields added by the global safe exception filter (T064).
      type: object
      required: [statusCode, message]
      properties:
        statusCode: { type: integer }
        message: { oneOf: [{type: string}, {type: array, items: {type: string}}] }
        error: { type: string, description: Existing Nest reason phrase }
        code: { type: string }
        correlationId: { type: string }
        fields: { type: object, additionalProperties: {type: array, items: {type: string}} }
...
    ParseOutcome:
      description: >-
        POST /email-messages/{id}/parse result (as built, T045). Either a transaction outcome or a sanitized failure.
        A message already linked to a transaction returns created false. A certain repeat (the same bank
        transaction code with the same amount, direction, and currency) returns the existing transaction with
        duplicate true and creates nothing. Any other match creates a row flagged as a suspected duplicate
        (isDuplicate, excluded from totals until cleared) and returns suspectedDuplicate true (EMAIL-008).
      oneOf:
        - type: object
          required: [transactionId, created]
          properties:
            transactionId: { type: string }
            created: { type: boolean }
            duplicate: { type: boolean, const: true, description: Present only for a certain cross-message duplicate }
            suspectedDuplicate: { type: boolean, const: true, description: Present only when a flagged suspected duplicate was created }
        - type: object
          required: [parserRun, created]
          properties:
            created: { const: false }
            parserRun: { $ref: '#/components/schemas/ParserRun' }
    ParserRun:
      description: >-
        Sanitized parse attempt (EMAIL-010, EMAIL-011). errorMessage is a code with field names only, for example
        "INVALID_AMOUNT: amount", "MISSING_REQUIRED_FIELDS: amount, transaction_time", "AMBIGUOUS_VALUE: amount",
        "AMBIGUOUS_DIRECTION: direction", "NO_TEMPLATE_MATCHED".
        extractedPayload holds per-field match status only (MATCHED, FALLBACK, MISSING), never captured text.
        normalizedPayload holds the deduplication decision only: { deduplication: { strategy, outcome: CREATED |
        DUPLICATE | SUSPECTED_DUPLICATE, duplicateOfTransactionId?, sameFacts? } }. Legacy rows written before T045
        may still hold captured values.
      type: object
      properties:
        id: { type: string }
        emailMessageId: { type: string }
        parserTemplateId: { type: [string, 'null'] }
        status: { enum: [SUCCESS, FAILED, PARTIAL] }
        confidenceScore: { type: [number, 'null'] }
        extractedPayload: { type: object }
        normalizedPayload: { type: object }
        errorMessage: { type: [string, 'null'] }
        createdTransactionId: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
...
    Alert:
      description: Existing alert fields plus DB-M4 lifecycle; read state (isRead/readAt) is independent of status.
      type: object
      required: [id, type, severity, title, message, isRead, status, triggeredAt, createdAt]
      properties:
        id: { type: string }
        type: { enum: [BUDGET_THRESHOLD, LARGE_TRANSACTION, CATEGORY_SHIFT, GOAL_RISK, CASHFLOW_RISK, PARSER_ISSUE, SYSTEM] }
        severity: { enum: [INFO, WARNING, CRITICAL] }
        title: { type: string }
        message: { type: string }
        resourceType: { type: [string, 'null'] }
        resourceId: { type: [string, 'null'] }
        isRead: { type: boolean }
        readAt: { type: [string, 'null'], format: date-time }
        status: { enum: [ACTIVE, DISMISSED, RESOLVED] }
        conditionKey: { type: [string, 'null'], description: null for legacy and user-authored alerts }
        thresholdValue: { type: [number, 'null'] }
        observedValue: { type: [number, 'null'] }
        periodStart: { type: [string, 'null'], format: date-time }
        periodEnd: { type: [string, 'null'], format: date-time }
        triggeredAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        resolutionReason: { type: [string, 'null'] }
        dismissedAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        emailDelivery: { oneOf: [{ $ref: '#/components/schemas/AlertDelivery' }, { type: 'null' }], description: 'null for legacy and user-authored alerts (no delivery row)' }
    AlertDelivery:
      type: object
      required: [channel, status, attemptCount]
      properties:
        channel: { enum: [EMAIL] }
        status: { enum: [PENDING, SENT, SKIPPED, FAILED] }
        skipReason: { type: [string, 'null'], enum: [NOT_CRITICAL, EMAIL_DISABLED, NOTIFICATIONS_DISABLED, TRANSPORT_DISABLED, null] }
        attemptCount: { type: integer, minimum: 0, maximum: 3 }
        lastAttemptAt: { type: [string, 'null'], format: date-time }
        sentAt: { type: [string, 'null'], format: date-time }
        failureCode: { type: [string, 'null'], description: 'e.g. TIMEOUT, AUTH, REJECTED, INTERRUPTED' }
    AlertSetting:
      description: Existing per-type setting; emailEnabled default becomes false for every type.
      type: object
      required: [type, inAppEnabled, emailEnabled]
      properties:
        type: { enum: [BUDGET_THRESHOLD, LARGE_TRANSACTION, CATEGORY_SHIFT, GOAL_RISK, CASHFLOW_RISK, PARSER_ISSUE, SYSTEM] }
        inAppEnabled: { type: boolean, default: true }
        emailEnabled: { type: boolean, default: false }
        threshold: { type: [number, 'null'], exclusiveMinimum: 0, description: 'LARGE_TRANSACTION amount in user base currency; null = default 5,000,000 when base currency is VND, otherwise inactive. As built: a write of 0 or less returns 400 for every type; null on PATCH clears it; a change never re-evaluates existing transactions.' }
        emailAvailable: { type: boolean, description: Additive; false when EMAIL_TRANSPORT=disabled }
...
    ClassificationProvenance:
      description: >-
        Additive transaction fields (T050): who decided the current category. classificationSource is MANUAL
        (locked), USER_RULE, SYSTEM_RULE, FALLBACK (no rule matched; categoryId null, the item needs a category),
        or UNKNOWN (automatic classification off, or a legacy row). The legacy enum values RULE, ML, LLM, and SYSTEM
        are never written (CLASS-007).
      type: object
      properties:
        classificationSource: { enum: [MANUAL, USER_RULE, SYSTEM_RULE, FALLBACK, UNKNOWN] }
        classificationRuleId: { type: [string, 'null'], description: 'Winning rule; null for MANUAL, FALLBACK, UNKNOWN, or a deleted rule' }
        classificationConfidence: { type: [number, 'null'], description: 1 when a category is set by the owner or a rule; null otherwise }
        classifiedAt: { type: [string, 'null'], format: date-time }
    DecisionExplanation:
      description: >-
        Why the winner won (CLASS-003): every matching eligible rule in rank order (user before system, then
        priority descending, createdAt ascending, id ascending), the tie-break between the winner and the runner-up,
        and whether the matches pointed to more than one category. Ids and codes only, never text (SEC-006).
      type: object
      required: [candidates, winnerRuleId, runnerUpRuleId, tieBreak, conflict]
      properties:
        candidates:
          type: array
          items:
            type: object
            required: [ruleId, scope, priority, createdAt, categoryId]
            properties:
              ruleId: { type: string }
              scope: { enum: [USER, SYSTEM] }
              priority: { type: integer }
              createdAt: { type: string, format: date-time }
              categoryId: { type: string }
        winnerRuleId: { type: [string, 'null'] }
        runnerUpRuleId: { type: [string, 'null'] }
        tieBreak: { enum: [SCOPE, PRIORITY, CREATED_AT, ID, null] }
        conflict: { type: boolean }
    CategoryEvent:
      description: One append-only category history event (CLASS-005), oldest first by sequence.
      type: object
      required: [id, sequence, createdAt, source, trigger, reason, actorType, actorUserId, merchantRuleId, previousCategory, newCategory, explanation]
      properties:
        id: { type: string }
        sequence: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        source: { enum: [MANUAL, USER_RULE, SYSTEM_RULE, FALLBACK] }
        trigger: { enum: [CREATE, IMPORT, MANUAL_CORRECTION, EXPLICIT_RECLASSIFY, AUTOMATIC_RERUN] }
        reason: { enum: [MANUAL_SET, MANUAL_CLEARED, RULE_MATCHED, NO_RULE_MATCHED] }
        actorType: { enum: [USER, SYSTEM] }
        actorUserId: { type: [string, 'null'] }
        merchantRuleId: { type: [string, 'null'], description: Kept after the rule is deleted }
        previousCategory:
          oneOf:
            - { type: 'null' }
            - { type: object, properties: { id: { type: string }, name: { type: [string, 'null'], description: 'Null when the owner can no longer read the category' } } }
        newCategory: { $ref: '#/components/schemas/CategoryEvent/properties/previousCategory' }
        explanation: { oneOf: [{ type: 'null' }, { $ref: '#/components/schemas/DecisionExplanation' }] }
...
    CurrencyTotals:
      description: Eligible totals of one currency (data-model.md "Financial period policy"); money is never summed across currencies.
      type: object
      required: [currency, income, expense, netCashflow, transactionCount]
      properties:
        currency: { type: string }
        income: { type: number }
        expense: { type: number }
        netCashflow: { type: number }
        transactionCount: { type: integer, description: Eligible records of every direction, transfers and adjustments included }
    TotalsSummary:
      description: >-
        Base-currency totals plus every currency group. income, expense, and netCashflow are the account base
        currency (named by currency); transactionCount spans every currency; currencies lists each group, sorted by
        code, and is empty when nothing is eligible.
      allOf:
        - $ref: '#/components/schemas/CurrencyTotals'
        - type: object
          required: [currencies]
          properties:
            currencies: { type: array, items: { $ref: '#/components/schemas/CurrencyTotals' } }
...
  responses:
    Unauthorized: { description: 'Missing, expired, or invalid session', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: Authenticated but insufficient role, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    OwnerSafeNotFound: { description: 'Absent, not owned, or malformed id (non-disclosing; administrators get the same outcome for private resources). A URL containing an encoded NUL (%00) gets this 404 before authentication; an id with undecodable percent-encoding (for example %FF) gets 400 before authentication instead', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    ValidationError: { description: Field-level validation failure, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    EmptyOk:
      description: Enveloped success with no specific payload
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
security:
  - cookieAuth: []
paths:
...
  /parser-templates:
    post:
      summary: Administrator creates a global parser template (SEC-003); reads (GET) stay available to authenticated users
      responses:
        '201': { description: Created template (enveloped) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /parser-templates/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      summary: Administrator updates a global parser template (SEC-003)
      responses:
        '200': { description: Updated template (enveloped) }
        '403': { $ref: '#/components/responses/Forbidden' }
  /transactions:
    get:
      summary: >-
        Existing owner list (never soft-deleted rows) plus additive totals: the eligible totals of every row matching the
        same filters, not only the page. With the same month and no other filter they equal the dashboard overview.
      parameters:
        - { in: query, name: month, schema: { type: string, pattern: '^\d{4}-(0[1-9]|1[0-2])$' }, description: 'Additive; a user month. Cannot be combined with from/to (400)' }
        - { in: query, name: from, schema: { type: string, format: date-time }, description: 'Inclusive; needs a UTC offset; must not be after to' }
        - { in: query, name: to, schema: { type: string, format: date-time }, description: 'Inclusive; needs a UTC offset' }
        - { in: query, name: search, schema: { type: string, maxLength: 200 } }
      responses:
        '200':
          description: 'Enveloped { data, total, page, limit, totals }'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        type: object
                        properties:
                          data: { type: array, items: { type: object } }
                          total: { type: integer }
                          page: { type: integer }
                          limit: { type: integer }
                          totals: { $ref: '#/components/schemas/TotalsSummary' }
        '400': { $ref: '#/components/responses/ValidationError' }
    post:
      summary: Existing owner create; body rules in TransactionWriteRules
      responses:
        '201': { description: Created transaction (enveloped) }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
...
  /transactions/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      summary: Existing owner update; body rules in TransactionWriteRules (no DELETED status, no provenance fields)
      responses:
        '200': { description: Updated transaction (enveloped) }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
    delete:
      summary: Existing owner soft delete, the only way to set DELETED (audited TRANSACTION_DELETED)
      responses:
        '200': { description: 'Enveloped { id }' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
...
  /email-messages/{id}/parse:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      summary: >-
        Parse one owned message again from its stored metadata (legacy rows may still carry a snippet). The raw body is
        never stored, so a message with neither is refused with 409 EMAIL_BODY_UNAVAILABLE and nothing is recorded.
      responses:
        '201':
          description: Enveloped ParseOutcome
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/ParseOutcome' } }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
        '409': { description: 'No stored body and none supplied; code EMAIL_BODY_UNAVAILABLE', content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  /transactions/{id}/category:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      summary: Apply authoritative manual category correction
      description: >-
        TX-004, CLASS-005: the category (or null to clear it) becomes a MANUAL decision that automatic runs never
        overwrite. The row is locked, updated, and given a MANUAL_CORRECTION event in one database transaction; the
        audit TRANSACTION_CATEGORY_CORRECTED is written when the category changes. Repeating the current manual
        choice writes nothing (decision.eventId null). categoryId is required (an empty body gets 400).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [categoryId]
              properties:
                categoryId: { type: [string, 'null'], minLength: 1 }
      responses:
        '200':
          description: The updated transaction with the decision (enveloped)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/ClassificationProvenance'
                          - type: object
                            properties:
                              decision:
                                type: object
                                required: [source, reason, eventId]
                                properties:
                                  source: { const: MANUAL }
                                  reason: { enum: [MANUAL_SET, MANUAL_CLEARED] }
                                  eventId: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
  /transactions/{id}/reclassify:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    post:
      summary: Explicit deterministic reclassification, releasing a manual category
      description: >-
        CLASS-006: applies the current rules to the transaction's current facts and replaces a MANUAL category
        (clients must warn first). Always appends an EXPLICIT_RECLASSIFY event and the audit TRANSACTION_RECLASSIFIED,
        even when the outcome is unchanged. The body must be empty (any property gets 400). Answers 200.
      requestBody:
        content: { application/json: { schema: { type: object, additionalProperties: false } } }
      responses:
        '200':
          description: The updated transaction with the decision (enveloped)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties:
                      data:
                        allOf:
                          - $ref: '#/components/schemas/ClassificationProvenance'
                          - type: object
                            properties:
                              decision:
                                type: object
                                required: [source, categoryId, ruleId, reason, explanation, eventId]
                                properties:
                                  source: { enum: [USER_RULE, SYSTEM_RULE, FALLBACK] }
                                  categoryId: { type: [string, 'null'] }
                                  ruleId: { type: [string, 'null'] }
                                  reason: { enum: [RULE_MATCHED, NO_RULE_MATCHED] }
                                  explanation: { $ref: '#/components/schemas/DecisionExplanation' }
                                  eventId: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
  /transactions/{id}/category-history:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      summary: The owner's append-only category history, oldest first
      description: >-
        Rows created before DB-M3 have no events (no history is backfilled). A soft-deleted transaction gets 404; its
        events are kept.
      responses:
        '200':
          description: Enveloped array of CategoryEvent
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { type: array, items: { $ref: '#/components/schemas/CategoryEvent' } } }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
...
  /alerts:
    get:
      parameters:
        - { in: query, name: status, schema: { enum: [ACTIVE, DISMISSED, RESOLVED] } }
        - { in: query, name: isRead, schema: { type: boolean }, description: 'Only "true" and "false" are accepted (400 otherwise). The existing `read` name is still accepted (as built, T068).' }
      responses:
        '200':
          description: Owner alerts (enveloped; existing pagination fields retained)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Alert' } } } } }
    post:
      summary: Existing user-authored alert; conditionKey null, never evaluated/resolved, in-app only with no delivery row (retention pending decision)
      responses:
        '201': { $ref: '#/components/responses/EmptyOk' }
...
  /alerts/{id}/read:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      summary: Existing; changes only isRead/readAt, never status
      responses:
        '200': { $ref: '#/components/responses/EmptyOk' }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
...
  /alerts/{id}/dismiss:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    patch:
      summary: ACTIVE -> DISMISSED; sets dismissedAt and marks read (an earlier readAt is kept); dismissing a DISMISSED alert returns it unchanged; RESOLVED alerts return 409 with code ALERT_RESOLVED
      responses:
        '200':
          description: Owned alert dismissed (enveloped)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { $ref: '#/components/schemas/Alert' } }
        '404': { $ref: '#/components/responses/OwnerSafeNotFound' }
        '409': { description: Alert already resolved, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /alerts/settings:
    get:
      responses:
        '200':
          description: Per-type settings (enveloped)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - properties: { data: { type: array, items: { $ref: '#/components/schemas/AlertSetting' } } }
    patch:
      requestBody: { content: { application/json: { schema: { $ref: '#/components/schemas/AlertSetting' } } } }
      responses:
        '200': { $ref: '#/components/responses/EmptyOk' }
        '400': { $ref: '#/components/responses/ValidationError' }
...